% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{calcul_distribution_semaine}
\alias{calcul_distribution_semaine}
\title{Calculer la distribution des trajets sur la semaine}
\usage{
calcul_distribution_semaine(trajet)
}
\arguments{
\item{trajet}{Un data.frame de trajets velo.}
}
\value{
Un data.frame contenant les jours de la semaine
et le nombre total de trajets.
}
\description{
Calcule le nombre total de trajets pour chaque jour de la semaine.
}
