% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{compter_nombre_trajets}
\alias{compter_nombre_trajets}
\title{Compter le nombre total de trajets}
\usage{
compter_nombre_trajets(trajet)
}
\arguments{
\item{trajet}{Un data.frame de trajets velo.}
}
\value{
Un nombre correspondant au total des trajets.
}
\description{
Calcule la somme de la colonne \code{Total}.
}
