% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-df_velo.R
\docType{data}
\name{df_velo}
\alias{df_velo}
\title{Trajets velo a Nantes pendant la Toussaint 2025}
\format{
A data.frame with several variables:
\describe{
\item{Jour}{Date du comptage.}
\item{Jour de la semaine}{Jour de la semaine code ou libelle.}
\item{Numéro de boucle}{Identifiant de la boucle de comptage.}
\item{Boucle de comptage}{Nom ou libelle de la boucle.}
\item{Total}{Nombre total de passages enregistres.}
\item{Probabilité de présence d'anomalies}{Indicateur d'anomalie.}
}
}
\source{
Nantes Metropole
}
\usage{
df_velo
}
\description{
Jeu de donnees contenant les comptages de trajets velo
a Nantes Metropole pendant les vacances de la Toussaint 2025.
}
\keyword{datasets}
