% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{filtrer_trajet}
\alias{filtrer_trajet}
\title{Filtrer un jeu de donnees par numero de boucle}
\usage{
filtrer_trajet(trajet, boucle)
}
\arguments{
\item{trajet}{Un data.frame de trajets velo.}

\item{boucle}{Un vecteur de numeros de boucle a conserver.}
}
\value{
Un data.frame filtre.
}
\description{
Filtrer un jeu de donnees par numero de boucle
}
