% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{plot_distribution_semaine}
\alias{plot_distribution_semaine}
\title{Tracer la distribution hebdomadaire des trajets}
\usage{
plot_distribution_semaine(trajet)
}
\arguments{
\item{trajet}{Un data.frame de trajets velo.}
}
\value{
Un graphique ggplot.
}
\description{
Filtre les anomalies, calcule la distribution des trajets
par jour de la semaine et produit un diagramme en colonnes.
}
