% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{trouver_trajet_max}
\alias{trouver_trajet_max}
\title{Trouver le trajet avec le maximum de passages}
\usage{
trouver_trajet_max(trajet)
}
\arguments{
\item{trajet}{Un data.frame de trajets velo.}
}
\value{
Un data.frame d'une ligne contenant la boucle,
le jour, le total, la moyenne du jour et la moyenne de la boucle.
}
\description{
Filtre les anomalies puis identifie la paire boucle-jour
avec le plus grand nombre de passages.
}
